Moved Camera folder to Script folder
