+++ FortressForge/Assets/Scripts/FortressForge.asmdef
@@ -1,7 +1,9 @@
 {
     "name": "FortressForge",
     "rootNamespace": "FortressForge",
-    "references": [],
+    "references": [
+        "GUID:75469ad4d38634e559750d17036d5f7c"
+    ],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
